+++ src/tests/services/fate.tests.ts
@@ -1,6 +1,12 @@
-import { FATE } from '../../defaults';
-import { mapResultToLeader, prepareExtendedFateRoll, rollFateDices, translateToFate } from '../../services';
-import { FateLeader } from '../../types';
+import { FATE, NO_DICE_FOUND_ERROR, ROLLS_RESULTS_FONTS } from '../../defaults';
+import {
+    mapFateToDice,
+    mapResultToLeader,
+    prepareExtendedFateRoll,
+    rollFateDices,
+    translateToFate,
+} from '../../services/fate.service';
+import { DiceTypes, FateLeader } from '../../types';
 import { FATE_ROLL, FATE_ROLL_EXTENDED_NO_MOD_MOCK, FATE_ROLL_EXTENDED_WITH_MOD_MOCK } from '../models/mocks';
 import * as tools from '../../services/rolls.service';
 
@@ -59,3 +65,25 @@ describe('rollFateDices', () => {
         expect(result).toEqual(expectedResult);
     });
 });
+
+describe('mapFateToDice', () => {
+    test('Should map Fate result to dice', () => {
+        const roll = -1;
+        const expectedResult = ROLLS_RESULTS_FONTS[DiceTypes.FATE][0];
+
+        const result = mapFateToDice(roll);
+        expect(result).toEqual(expectedResult);
+    });
+
+    test('Should throw error if there is no possibility to map result', () => {
+        const roll = 66;
+        const expectedError = NO_DICE_FOUND_ERROR;
+
+        const callback = () => {
+            mapFateToDice(roll);
+        };
+        expect(callback).toThrow(expectedError);
+    });
+});
+
+export {};
